Refactored backed, created push request
Refactored the backend to allow for my custom data types, replacing the existing TODO elements. Created the push functionality... sort of. Attempted the pull requests but ultimately the project does not work.

I was a little too ambitious for the time I had to work on the project. Redoing the backend with GoLang and Docker, two things I don't know well, was NOT for the faint of heart.